Add files via upload
Information gotten from the data includes; total number of employees both active and terminated, age range of employees, average salary of workers. I also looked at the best route to employ, the ration of the gender in the company, top reasons why the employees left,  percentage of people that voluntarily  left to people who were sacked and people who is still present in the company.